Safari-609.1.11.1/LayoutTests/platform/ios-wk2/imported/w3c/web-platform-tests/css/css-position/position-absolute-percentage-height-expected.txt


Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet.

FAIL #target height matches containing block height, and target parent width matches #target width assert_equals: expected 160 but got 0
FAIL #target height matches containing block height, and target parent width matches #target width after resize assert_equals: expected 240 but got 0
